feat: Testing component inputs in edit mode
- moved PropsEditor into a separate component
