+++ frontend/components.d.ts
@@ -31,7 +31,7 @@ declare module 'vue' {
     ComponentInterface: typeof import('./src/components/ComponentInterface.vue')['default']
     ComponentLayers: typeof import('./src/components/ComponentLayers.vue')['default']
     ComponentPanel: typeof import('./src/components/ComponentPanel.vue')['default']
-    ComponentProps: typeof import('./src/components/ComponentProps.vue')['default']
+    ComponentProperties: typeof import('./src/components/ComponentProperties.vue')['default']
     ComponentStyles: typeof import('./src/components/ComponentStyles.vue')['default']
     Container: typeof import('./src/components/AppLayout/Container.vue')['default']
     ContextMenu: typeof import('./src/components/ContextMenu.vue')['default']
@@ -65,6 +65,7 @@ declare module 'vue' {
     PagesPanel: typeof import('./src/components/PagesPanel.vue')['default']
     PanelResizer: typeof import('./src/components/PanelResizer.vue')['default']
     PlacementControl: typeof import('./src/components/PlacementControl.vue')['default']
+    PropsEditor: typeof import('./src/components/PropsEditor.vue')['default']
     ProxyDialog: typeof import('./src/components/ProxyComponents/ProxyDialog.vue')['default']
     Repeater: typeof import('./src/components/AppLayout/Repeater.vue')['default']
     RepeaterContextProvider: typeof import('./src/components/AppLayout/RepeaterContextProvider.vue')['default']
